refactor: MVC, Design Patterns, Pessoa creation"
- Use layers to improve logic: Model, View, Controller
- Use Design Patterns - following DRY: Factory and Strategy
- Pessoa can now be properly created as either an Adotante, Funcionario or Tutor.

+++ src/model/Funcionario.java
@@ -5,7 +5,7 @@ public class Funcionario {
     private long numeroIdentificacao;
     private String dataContratacao;
     private String cargo;
-    private float salario;
+    private double salario;
     private String departamento;
 
     public Funcionario() {}
@@ -45,11 +45,11 @@ public class Funcionario {
         this.cargo = cargo;
     }
 
-    public float getSalario() {
+    public double getSalario() {
         return salario;
     }
 
-    public void setSalario(float salario) {
+    public void setSalario(double salario) {
         this.salario = salario;
     }
